+++ js/main.js
@@ -1,6 +1,6 @@
 import {getArrayOfAds, MAX_ADS_AMOUNT} from './data.js';
 import {drawAd} from './draw-ads.js';
-import {setPageActive} from './page-condition.js';
+import {setFormActive} from './form.js';
 getArrayOfAds(MAX_ADS_AMOUNT);
 drawAd();
-setPageActive();
+setFormActive();
